move bft AT to bft package
